requirements.txt
flask
python-dotenv
openai
pillow
